requirements.txt
mock
twill
django
south
setuptools
psycopg2
PIL
xmlbuilder
django-crontab
django-file-resubmit
django-cors-headers
numpy
django-tinymce
